Use a distribution to backsolve case count from fatalities.
Here, I use a gaussian distribution with sigma = 5.25 and mu = -19. This comes
from an early paper on the time course of COVID, published in the Lancet back
in March. In it, the authors estimate that in those hospitalized patients
who died of COVID-19, the mean time from infection to death was 18.5 days with
a mean-centered IQR of 7 days - this corresponds to a sigma of 5.25 when
approximating as a normal curve.